requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiogram==3.1.1
schedule==1.2.0
httpx==0.27.0
tenacity==8.2.3
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
